Refactor: Rename `Message.providers()` to `provider()`
The `Message.providers()` method has been renamed to `Message.provider()` for improved naming consistency.

This refactoring updates the primary method for accessing the list of `ContextProvider` objects within a `Message`. All internal call sites within the `Message` and `Messages` classes, as well as the test suite, have been updated to reflect this change.

The project version is also bumped to 0.1.48.